+++ src/custom-nodes/string-array-nodes/StringArrayCallsWrapper.ts
@@ -100,6 +100,7 @@ export class StringArrayCallsWrapper extends AbstractCustomNode {
             {
                 ...NO_ADDITIONAL_NODES_PRESET,
                 identifierNamesGenerator: this.options.identifierNamesGenerator,
+                identifiersDictionary: this.options.identifiersDictionary,
                 seed: this.options.seed
             }
         ).getObfuscatedCode();