+++ src/enums/generators/identifier-names-generators/IdentifierNamesGenerator.ts
@@ -1,4 +1,5 @@
 export enum IdentifierNamesGenerator {
+    DictionaryIdentifierNamesGenerator = 'dictionary',
     HexadecimalIdentifierNamesGenerator = 'hexadecimal',
     MangledIdentifierNamesGenerator = 'mangled'
 }