+++ test/functional-tests/javascript-obfuscator/fixtures/dictionary-identifiers.js
@@ -1,0 +1,5 @@
+(function () {
+    var test1 = 1;
+    var test2 = 2;
+    var test3 = 3;
+})();