+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 Change Log
 
+v0.20.0
+---
+* **Breaking:** dropped support of Node 8 because of end of maintenance support
+* **New option value:** `identifierNamesGenerator` now allows to set new `dictionary` identifier names generator
+* **New option:** `identifiersDictionary` sets identifiers dictionary for `identifierNamesGenerator: dictionary` option
+
+Thanks to our contributors!
+ * [adiantek](https://github.com/adiantek)
+
 v0.19.4
 ---
 * Fixed `reservedNames` option